removed old benchmarks and added vitest benchmarks for proxyMap
